Refactored manifest retrieval into Core so UI can use it
