+++ SirenOfShame.Uwp.Core/SirenOfShame.Uwp.Core.csproj
@@ -5,6 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="Newtonsoft.Json" Version="10.0.3" />
     <PackageReference Include="Polly" Version="5.7.0" />
   </ItemGroup>
