--- SirenOfShame.Uwp.Maintenance/Interfaces/ILog.cs
@@ -1,14 +1,0 @@
-﻿using System;
-using System.Threading.Tasks;
-
-namespace SirenOfShame.Uwp.Maintenance
-{
-    internal interface ILog
-    {
-        Task Error(string message);
-        Task Error(string message, Exception exception);
-        Task Warn(string message);
-        Task Info(string message);
-        Task Debug(string message);
-    }
-}
