+++ SirenOfShame.Uwp.Maintenance/SirenOfShame.Uwp.Maintenance.csproj
@@ -98,9 +98,7 @@
     <RestoreProjectStyle>PackageReference</RestoreProjectStyle>
   </PropertyGroup>
   <ItemGroup>
-    <Compile Include="Interfaces\ILog.cs" />
     <Compile Include="Log\MessageRelayLogger.cs" />
-    <Compile Include="Models\Bundle.cs" />
     <Compile Include="Services\BundleService.cs" />
     <Compile Include="Services\CertificatePinningHttpClientFactory.cs" />
     <Compile Include="Services\MessageRelayService.cs" />
@@ -150,6 +148,9 @@
       <Project>{7668e982-ca88-4fa0-8332-a1a2867cad0b}</Project>
       <Name>SirenOfShame.Uwp.Core</Name>
     </ProjectReference>
+  </ItemGroup>
+  <ItemGroup>
+    <Folder Include="Models\" />
   </ItemGroup>
   <PropertyGroup Condition=" '$(VisualStudioVersion)' == '' or '$(VisualStudioVersion)' &lt; '15.0' ">
     <VisualStudioVersion>15.0</VisualStudioVersion>
